Engraving: move reset commands from Score into Reset
Moves cmdResetAllPositions, cmdResetToDefaultLayout,
cmdResetTextStyleOverrides, cmdResetAllStyles and cmdResetMeasuresLayout
out of Score into a new static Reset class.
