+++ src/engraving/editing/cmd.cpp
@@ -124,34 +124,6 @@ static void resetElementPosition(EngravingItem* e)
     }
 }
 
-static void resetTextProperties(EngravingItem* e)
-{
-    if (e->generated() || !e->isTextBase()) {
-        return;
-    }
-
-    static const std::vector<Pid> TEXT_STYLE_TO_RESET {
-        Pid::FONT_FACE,
-        Pid::FONT_SIZE,
-        Pid::FONT_STYLE,
-        Pid::SIZE_SPATIUM_DEPENDENT,
-        Pid::FRAME_TYPE,
-        Pid::TEXT_LINE_SPACING,
-        Pid::FRAME_FG_COLOR,
-        Pid::FRAME_BG_COLOR,
-        Pid::FRAME_WIDTH,
-        Pid::FRAME_PADDING,
-        Pid::FRAME_ROUND,
-        Pid::ALIGN
-    };
-
-    for (Pid pid : TEXT_STYLE_TO_RESET) {
-        // TODO: use undoResetProperty: https://github.com/musescore/MuseScore/issues/16516
-        // But for now, we'll use resetPropety since undoResetProperty leads to various problems
-        e->resetProperty(pid);
-    }
-}
-
 //---------------------------------------------------------
 //   reset
 //---------------------------------------------------------
@@ -1869,207 +1841,6 @@ void Score::cmdAddStretch(double val)
     }
 }
 
-void Score::cmdResetToDefaultLayout()
-{
-    TRACEFUNC;
-
-    Transaction& tx = transactionManager()->currentOrDummyTransaction();
-
-    StyleIdSet dontResetTheseStyles {
-        Sid::lyricsPlacement,
-        Sid::repeatBarTips,
-        Sid::startBarlineSingle,
-        Sid::startBarlineMultiple,
-        Sid::dividerLeft,
-        Sid::dividerRightY,
-        Sid::useStraightNoteFlags,
-        Sid::mrNumberSeries,
-        Sid::mrNumberEveryXMeasures,
-        Sid::mrNumberSeriesWithParentheses,
-        Sid::oneMeasureRepeatShow1,
-        Sid::fourMeasureRepeatShowExtenders,
-        Sid::useWideBeams,
-        Sid::hairpinPlacement,
-        Sid::pedalPlacement,
-        Sid::trillPlacement,
-        Sid::vibratoPlacement,
-        Sid::harmonyPlacement,
-        Sid::romanNumeralPlacement,
-        Sid::nashvilleNumberPlacement,
-        Sid::harmonyVoiceLiteral,
-        Sid::harmonyVoicing,
-        Sid::harmonyDuration,
-        Sid::capoPosition,
-        Sid::fretNumPos,
-        Sid::fretPlacement,
-        Sid::fretStrings,
-        Sid::fretFrets,
-        Sid::fretNut,
-        Sid::fretOrientation,
-        Sid::showPageNumber,
-        Sid::showPageNumberOne,
-        Sid::pageNumberOddEven,
-        Sid::showMeasureNumber,
-        Sid::showMeasureNumberOne,
-        Sid::measureNumberInterval,
-        Sid::measureNumberSystem,
-        Sid::measureNumberPlacementMode,
-        Sid::genClef,
-        Sid::hideTabClefAfterFirst,
-        Sid::genKeysig,
-        Sid::genCourtesyTimesig,
-        Sid::genCourtesyKeysig,
-        Sid::genCourtesyClef,
-        Sid::swingRatio,
-        Sid::swingUnit,
-        Sid::chordSymbolSpelling,
-        Sid::automaticCapitalization,
-        Sid::lowerCaseMinorChords,
-        Sid::lowerCaseBassNotes,
-        Sid::allCapsNoteNames,
-        Sid::chordStyle,
-        Sid::chordsXmlFile,
-        Sid::chordDescriptionFile,
-        Sid::concertPitch,
-        Sid::createMultiMeasureRests,
-        Sid::minEmptyMeasures,
-        Sid::hideEmptyStaves,
-        Sid::dontHideStavesInFirstSystem,
-        Sid::enableIndentationOnFirstSystem,
-        Sid::firstSystemIndentationValue,
-        Sid::hideInstrumentNameIfOneInstrument,
-        Sid::gateTime,
-        Sid::tenutoGateTime,
-        Sid::staccatoGateTime,
-        Sid::slurGateTime,
-        Sid::sectionPause,
-        Sid::showHeader,
-        Sid::headerFirstPage,
-        Sid::headerOddEven,
-        Sid::evenHeaderL,
-        Sid::evenHeaderC,
-        Sid::evenHeaderR,
-        Sid::oddHeaderL,
-        Sid::oddHeaderC,
-        Sid::oddHeaderR,
-        Sid::showFooter,
-        Sid::footerFirstPage,
-        Sid::footerOddEven,
-        Sid::evenFooterL,
-        Sid::evenFooterC,
-        Sid::evenFooterR,
-        Sid::oddFooterL,
-        Sid::oddFooterC,
-        Sid::oddFooterR,
-        Sid::tupletOutOfStaff,
-        Sid::tupletDirection,
-        Sid::tupletBracketType,
-        Sid::dynamicsPlacement,
-        Sid::textLinePlacement,
-        Sid::harpPedalDiagramPlacement,
-        Sid::harpPedalTextDiagramPlacement,
-        Sid::expressionPlacement,
-        Sid::tupletNumberType,
-        Sid::mmRestShowMeasureNumberRange,
-        Sid::mmRestRangeBracketType,
-        Sid::mmRestRangeVPlacement,
-        Sid::staffTextPlacement,
-        Sid::repeatLeftPlacement,
-        Sid::repeatRightPlacement,
-        Sid::instrumentChangePlacement,
-        Sid::stickingPlacement,
-        Sid::letRingPlacement,
-        Sid::palmMutePlacement,
-        Sid::pageWidth,
-        Sid::pageHeight,
-        Sid::pagePrintableWidth,
-        Sid::pageEvenTopMargin,
-        Sid::pageEvenBottomMargin,
-        Sid::pageEvenLeftMargin,
-        Sid::pageOddTopMargin,
-        Sid::pageOddBottomMargin,
-        Sid::pageOddLeftMargin,
-        Sid::pageTwosided,
-        Sid::spatium,
-        Sid::concertPitch,
-        Sid::createMultiMeasureRests
-    };
-
-    auto resetPositionAndTextProperties = [](EngravingItem* e) {
-        resetElementPosition(e);
-        resetTextProperties(e);
-    };
-
-    cmdResetMeasuresLayout();
-    scanElements(resetPositionAndTextProperties);
-    cmdResetAllStyles(dontResetTheseStyles);
-    EditSystemLocks::undoRemoveAllLocks(tx, this);
-}
-
-void Score::cmdResetTextStyleOverrides()
-{
-    TRACEFUNC;
-
-    scanElements(resetTextProperties);
-}
-
-void Score::cmdResetAllStyles(const StyleIdSet& exceptTheseOnes)
-{
-    TRACEFUNC;
-
-    int beginIdx = int(Sid::NOSTYLE) + 1;
-    int endIdx = int(Sid::STYLES);
-
-    StyleIdSet stylesToReset;
-
-    for (int idx = beginIdx; idx < endIdx; idx++) {
-        Sid styleId = Sid(idx);
-        if (!muse::contains(exceptTheseOnes, styleId)) {
-            stylesToReset.insert(styleId);
-        }
-    }
-
-    score()->resetStyleValues(stylesToReset);
-}
-
-// Removes system/page breaks and spacers
-void Score::cmdResetMeasuresLayout()
-{
-    TRACEFUNC;
-
-    std::vector<EngravingItem*> itemsToRemove;
-
-    for (MeasureBase* mb = first(); mb; mb = mb->next()) {
-        if (mb->isMeasure()) {
-            mb->undoResetProperty(Pid::USER_STRETCH);
-
-            for (const MStaff* staff : toMeasure(mb)->mstaves()) {
-                if (Spacer* spacer = staff->vspacerDown()) {
-                    itemsToRemove.push_back(spacer);
-                }
-
-                if (Spacer* spacer = staff->vspacerUp()) {
-                    itemsToRemove.push_back(spacer);
-                }
-            }
-        }
-
-        if (!mb->lineBreak() && !mb->pageBreak()) {
-            continue;
-        }
-
-        for (EngravingItem* item : mb->el()) {
-            if (item->isLayoutBreak() && !toLayoutBreak(item)->isSectionBreak()) {
-                itemsToRemove.push_back(item);
-            }
-        }
-    }
-
-    for (EngravingItem* item : itemsToRemove) {
-        undoRemoveElement(item);
-    }
-}
-
 //---------------------------------------------------------
 //   cmdResetNoteAndRestGroupings
 //---------------------------------------------------------
@@ -2115,22 +1886,6 @@ static void resetBeamOffSet(EngravingItem* e)
     // Reset completely cross staff beams from MU1&2
     if (e->isBeam() && toBeam(e)->fullCross()) {
         e->reset();
-    }
-}
-
-//---------------------------------------------------------
-//   cmdResetAllPositions
-//---------------------------------------------------------
-void Score::cmdResetAllPositions(bool undoable)
-{
-    TRACEFUNC;
-
-    if (undoable) {
-        startCmd(TranslatableString("undoableAction", "Reset all positions"));
-    }
-    resetAutoplace();
-    if (undoable) {
-        endCmd();
     }
 }
