+++ src/engraving/tests/readwriteundoreset_tests.cpp
@@ -23,6 +23,10 @@
 
 #include "engraving/dom/masterscore.h"
 
+#include "engraving/editing/reset.h"
+#include "engraving/editing/transaction/transaction.h"
+
+#include "types/translatablestring.h"
 #include "utils/scorerw.h"
 #include "utils/scorecomp.h"
 
@@ -48,7 +52,9 @@ TEST_F(Engraving_ReadWriteUndoResetTests, testReadWriteResetPositions)
 
         MasterScore* score = ScoreRW::readScore(readFile);
         EXPECT_TRUE(score);
-        score->cmdResetAllPositions();
+        score->transactionManager()->transaction(muse::TranslatableString::untranslatable("Reset all positions"), [&](Transaction& tx) {
+            Reset::resetAllPositions(tx, score);
+        });
         score->undoRedo(/* undo */ true, nullptr);
         EXPECT_TRUE(ScoreComp::saveCompareScore(score, writeFile, readFile));
